+++ modules/segmentshortestpaths/core/segmtulipshortestpath.cpp
@@ -152,7 +152,7 @@ AnalysisResult SegmentTulipShortestPath::run(Communicator *) {
     }
 
     m_map.overrideDisplayedAttribute(-2); // <- override if it's already showing
-    m_map.setDisplayedAttribute(angle_col);
+    m_map.setDisplayedAttribute(static_cast<int>(angle_col));
 
     result.completed = true;
